feat(bootloader): Add ethernet; Update RUN_LINUX ifdef. - Add ethernet file transfers (only in fpga) to speed up boot time. - Update method to check if running Linux during runtime, without having   to recompile the bootloader. Related to: https://github.com/IObundle/iob-soc-opencryptolinux/commit/113f8de3e9c6d54cb3763e164503f48ed6cbc4dd#r135914395
Note, due to other macros, like `BAUD`, when re-building the system with
a different `RUN_LINUX` macro in simulation, the bootloader still needs to be
recompiled, since the simulation baud is different for RUN_LINUX=1 and 0.

However, for a given fpga board, the bootloader does not need to be
recompiled because the RUN_LINUX=1 and 0 macros have the same baud for
the FPGAs.